// src/interfaces/ISkillListImport.ts
export interface ISkillListImport {
    Name?: string;
    Attribute?: string;
    IsKnowledge?: boolean;
    Language?: boolean;
    AlwaysLanguage?: boolean;
    BaseParry?: boolean;

    name: string;
    attribute: string;
    is_knowledge: boolean;
    language: boolean;
    always_language: boolean;
    base_parry: boolean;
}